Moved functions that were not strictly attached to colly into a 'utils' file for better readability
